feat: move forVerticalInvertedIOS from default presets
